Voice İnput
In this version, voice input capabilities were added using google speech api. The robot waits for voice input, evaluates, and performs a movement. At this point, the movement performed by the robot is a dummy move. We will be working on adding more meaningful gestures in the commits that follow.